fix: moved table maker
